src/main/java/com/openclassrooms/PayMyBuddy/service/UserDetailsServiceImpl.java
package com.openclassrooms.PayMyBuddy.service;

import com.openclassrooms.PayMyBuddy.exception.UserNotFoundException;
import com.openclassrooms.PayMyBuddy.model.User;
import com.openclassrooms.PayMyBuddy.model.UserDetailsImpl;
import com.openclassrooms.PayMyBuddy.repository.UserRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

import java.sql.Timestamp;
import java.util.Calendar;

@Service
public class UserDetailsServiceImpl implements UserDetailsService {

    @Autowired
    private UserRepository userRepository;

    @Override
    public UserDetails loadUserByUsername(String userName) throws UsernameNotFoundException {
        User user = userRepository.findByEmailIgnoreCase(userName).orElseThrow(UserNotFoundException::new);
        Calendar calendar = Calendar.getInstance();
        user.setLastOnlineTime(new Timestamp(calendar.getTime().getTime()));
        userRepository.save(user);
        return new UserDetailsImpl(user);
    }
}
